backend/requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil >= 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.
Flask-Cors >= 3.0.8
pandas >= 1.3.4
tensorflow >= 2.9.1
scikit-learn >= 1.0.1
seaborn >= 0.11.2
pg8000 >= 1.29.0
sqlalchemy >= 1.3.0
google-cloud >= 0.30.0
cloud-sql-python-connector[pg8000]